+++ devcheck/__init__.py
@@ -1,0 +1,4 @@
+# ~/clientfactory/devcheck/__init__.py
+"""
+Utilities for checking development preparedness
+"""