+++ devcheck/core.py
@@ -1,0 +1,12 @@
+# ~/clientfactory/devcheck/core.py
+"""
+common utilities for other check scripts
+"""
+
+def getfilename(): pass
+
+def getlineno(): pass
+
+def getmethod(): pass
+
+def getclass(): pass