+++ devcheck/docstrings.py
@@ -1,0 +1,10 @@
+# ~/clientfactory/devcheck/docstrings.py
+"""
+check for empty/missing docstrings
+"""
+
+EMPTYVALS = {
+    """""",
+    """...""",
+    # or absence altogether
+}