+++ devcheck/notimpl.py
@@ -1,0 +1,4 @@
+# ~/clientfactory/devcheck/notimpl.py
+"""
+check if any methods are raising NotImplemented or NotImplementedError
+"""