+++ devcheck/shebangs.py
@@ -1,0 +1,4 @@
+# ~/clientfactory/devcheck/shebangs.py
+"""
+check for comments that start with '#!'
+"""